feat: move dylib code into seperate module
